refactor: remove legacy video js
