+++ src/index.ts
@@ -1,7 +1,8 @@
-import Video from './MyVideo';
+import Video from './Video';
 export { default as FilterType } from './lib/FilterType';
 export { default as VideoResizeMode } from './lib/VideoResizeMode';
 export { default as TextTrackType } from './lib/TextTrackType';
 export { default as DRMType } from './lib/DRMType';
-export { VideoRef } from './MyVideo';
+export { VideoRef } from './Video';
+
 export default Video;
